ai/openai/requirements.txt
ipython
pyyaml
httpx[socks]

openai
